% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_formula}
\alias{create_formula}
\title{Create dcast formula from two character vectors}
\usage{
create_formula(cols_lhs, cols_rhs)
}
\arguments{
\item{cols_lhs}{LHS columns}

\item{cols_rhs}{RHS columns}
}
\value{
formula
}
\description{
Create dcast formula from two character vectors
}
