% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_hash_from_data}
\alias{create_hash_from_data}
\title{Generate a SHA256 hash file from a downloaded WHO data file}
\usage{
create_hash_from_data(data)
}
\arguments{
\item{data}{A data object}
}
\value{
A character string of length 64 containing the SHA256 string
}
\description{
Generate a SHA256 hash file from a downloaded WHO data file
}
