% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_ind}
\alias{download_ind}
\title{Download file of indicator and read it into a data frame}
\usage{
download_ind(ind)
}
\arguments{
\item{ind}{A valid indicator of the GHO API.}
}
\value{
Tibble with data of indicator
}
\description{
Download file of indicator and read it into a data frame
}
