% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ind_in_cache}
\alias{ind_in_cache}
\title{Check if indicator data is already downloaded and in "cache"}
\usage{
ind_in_cache(ind)
}
\arguments{
\item{ind}{A valid indicator of the GHO API.}
}
\value{
TRUE if indicator is in cache, FALSE if not.
}
\description{
Check if indicator data is already downloaded and in "cache"
}
\examples{
ind_in_cache("RSUD_30")
}
