% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ind_to_last_update}
\alias{ind_to_last_update}
\title{Retrieve the last update date of an indicator from WHO}
\usage{
ind_to_last_update(ind)
}
\arguments{
\item{ind}{A valid indicator of the GHO API.}
}
\value{
A Date object containing the last update
}
\description{
Retrieve the last update date of an indicator from WHO
}
