% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ind_to_url}
\alias{ind_to_url}
\title{Indicator value to valid API url}
\usage{
ind_to_url(ind)
}
\arguments{
\item{ind}{An valid indicator of the GHO API.}
}
\value{
A valid GHO API url in string format.
}
\description{
Indicator value to valid API url
}
\examples{
ind_to_url("RSUD_30")
}
