% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_to_NA_list}
\alias{null_to_NA_list}
\title{Turn all NULL elements of list to NA.}
\usage{
null_to_NA_list(list)
}
\arguments{
\item{list}{A list object}
}
\value{
A list object with NULL elements replaced by NA
}
\description{
Turn all NULL elements of list to NA.
}
