% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_indicators.R
\name{read_GHO_data}
\alias{read_GHO_data}
\title{Load all data for the provided indicator.}
\usage{
read_GHO_data(indicator)
}
\arguments{
\item{indicator}{A valid WHO indicator. Check show_GHO_indicators() for options.}
}
\value{
A data frame containing all data of the WHO indicator.
}
\description{
Load all data for the provided indicator.
}
\examples{
read_GHO_data("WHOSIS_000001")
}
