% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{separate_url}
\alias{separate_url}
\title{Separate parts of the API url for further analysis}
\usage{
separate_url(url)
}
\arguments{
\item{url}{A GHO API call.}
}
\value{
A matrix with the protocol, host, api, indicator and possible filter.
}
\description{
Separate parts of the API url for further analysis
}
