% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_indicators.R
\name{show_GHO_indicators}
\alias{show_GHO_indicators}
\title{Load table with all available indicators for the GHO API.}
\usage{
show_GHO_indicators()
}
\value{
A data frame with all available indicators.
}
\description{
In the background we load https://ghoapi.azureedge.net/api/Indicator. This
is a JSON file with all indicators which are available in the WHO database.
}
