% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_indicators.R
\name{tidy_data}
\alias{tidy_data}
\title{Transform table according to tidy philosophy}
\usage{
tidy_data(data)
}
\arguments{
\item{data}{A raw WHO API dataset}
}
\value{
A tidy-transformed WHO dataset
}
\description{
All data tables requested by the WHO API contain a SpatialDim, TimeDim,
Dim1, Dim2, Dim3 and DataSourceDim dimension. These dimensions have 
corresponding units, e.g. the type of the TimeDim column is typically
a year.
}
\details{
We can transform the dimension and its corresponding dimension type to one 
column by setting the values of DimType to the column name and letting 
Dim be the values. Note that this sometimes results into multiple columns if 
if the DimType reports multiple units like years and months.
}
