% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ind_to_json}
\alias{ind_to_json}
\title{Indicator value to local JSON file}
\usage{
ind_to_json(ind)
}
\arguments{
\item{ind}{An valid indicator of the GHO API.}
}
\value{

}
\description{
Indicator value to local JSON file
}
